Add files via upload
marlin MP10 source code. It appears to be set up for the MP10 currently but has commented out sections for other Monoprice printers. I saw that it was missing on Monoprice3D's github and I was able to get it from Monoprice directly. I will keep it up here until they publish it themselves. As with all 3d Printer firmware please research before you change any firmware and be careful as mistakes can break the printer or cause safety issues. All responsibility for problems or damage is on the user. I do not work for Monoprice and make no claims for the company itself. Changing your firmware may result in damage and/or loss of warranty. 

All that aside happy printing and I hope these files are useful.